Runner: replace orAnother() with orElseGet()
Refactor usage of method org.openjdk.jmh.util.Optional#orAnother in
class Runner using new method orElseGet().  This is part of preparation
for a possible future drop-in replacement of usages of class
org.openjdk.jmh.util.Optional with class java.util.Optional.

In Java 8, the minimal version that JMH supports at the moment, class
java.util.Optional doesn't have an analogue to JMH's method orAnother().
A similar method (taking a Supplier of Optional as a parameter), which
could be used instead was added to java.util.Optional in Java 9, see
https://bugs.openjdk.org/browse/JDK-8080418.